finished tests

+++ app/Http/Controllers/HomeController.php
@@ -42,7 +42,9 @@ class HomeController extends Controller
 
         $projects = Project::orderBy('created_at', 'desc')->get();
 
-        return ProjectResource::collection($projects);
+        return ProjectResource::collection($projects)
+            ->response()
+            ->setStatusCode(201);
     }
 
     public function updateProject(Request $request){
@@ -67,7 +69,16 @@ class HomeController extends Controller
     }
 
     public function deleteProject(Request $request){
-        Project::find($request->project_id)->delete();
+
+        $token = $request->bearerToken();
+        $user = User::where('api_token', $token)->firstOrFail();
+        $project = Project::find($request->project_id);
+        if($project->user_id != $user->id){
+            return response()->json([
+                'data' => 'Nie możesz usuwać nie swój projekt'
+            ], 401);
+        };
+        $project->delete();
         Task::where('project_id', $request->project_id)->delete();
         $projects = Project::orderBy('created_at', 'desc')->get();
         return ProjectResource::collection($projects);
@@ -90,10 +101,22 @@ class HomeController extends Controller
 
         $tasks = Task::orderBy('created_at', 'desc')->get();
 
-        return TaskResource::collection($tasks);
+        return TaskResource::collection($tasks)
+                ->response()
+                ->setStatusCode(201);
     }
 
     public function updateTask(Request $request){
+
+        $token = $request->bearerToken();
+        $user = User::where('api_token', $token)->firstOrFail();
+        $projectsIds = $user->projects->pluck('id')->toArray();
+        $tasksId = Task::whereIn('project_id', $projectsIds)->pluck('id')->toArray();
+        if(!in_array($request->task_id, $tasksId)){
+            return response()->json([
+                'data' => 'Nie możesz aktualizować nie swój task'
+            ], 401);
+        };
 
         $task = Task::find($request->task_id);
         $task->title = $request->title;
@@ -106,6 +129,16 @@ class HomeController extends Controller
     }
 
     public function deleteTask(Request $request){
+
+        $token = $request->bearerToken();
+        $user = User::where('api_token', $token)->firstOrFail();
+        $projectsIds = $user->projects->pluck('id')->toArray();
+        $tasksId = Task::whereIn('project_id', $projectsIds)->pluck('id')->toArray();
+        if(!in_array($request->task_id, $tasksId)){
+            return response()->json([
+                'data' => 'Nie możesz usuwać nie swój task'
+            ], 401);
+        };
         Task::find($request->task_id)->delete();
         $tasks = Task::orderBy('created_at', 'desc')->get();
         return TaskResource::collection($tasks);
